Toolbar warning + load order fix + warning changes
When in darktide profile and toolbar is not activated makes a popup to activate it
Fix : when installing a mod now checks if it's managed by vortex or not.
Made some changes to how we made warnings, now it's 1 function that can take 3 args and warns can now never have the same ID so each warn call will always appear.